requirements.txt
Flask
requests
python-dotenv
Flask-Caching
Flask-Limiter
pytest
